debug: docker

+++ ansys/heart/preprocessor/mesh/mesher.py
@@ -465,12 +465,14 @@ def mesh_from_non_manifold_input_model(
     if _uses_container:
         config_dict = pyfluent.launch_fluent(start_container=True, dry_run=True)
         LOGGER.debug(f"Config dict:{config_dict}")
-        config_dict["host_mount_path"] = work_dir_meshing
+        # config_dict["host_mount_path"] = work_dir_meshing
+        config_dict = {}
         # work_dir_meshing = config_dict["working_dir"]
-    
-    if os.path.isdir(work_dir_meshing):
-        shutil.rmtree(work_dir_meshing)
-    os.makedirs(work_dir_meshing)
+        work_dir_meshing = pyfluent.EXAMPLES_PATH
+    if not _uses_container:
+        if os.path.isdir(work_dir_meshing):
+            shutil.rmtree(work_dir_meshing)
+        os.makedirs(work_dir_meshing)
 
     path_to_output_old = path_to_output
     path_to_output = os.path.join(work_dir_meshing, "volume-mesh.msh.h5")
